Life is meaningless.  Everything ends

+++ core/main.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 
 int main() {
-    btree<int> b_tree(5);
+    btree<double> b_tree(5);
 	b_tree.insert(1);
